Correção Exceptions SQL

+++ adabook/src/main/java/br/com/ada/adabook/service/AuthorServiceImpl.java
@@ -1,12 +1,15 @@
 package br.com.ada.adabook.service;
 
 import br.com.ada.adabook.domain.Author;
-import br.com.ada.adabook.dto.author.AuthorDescriptionDTO;
+import br.com.ada.adabook.domain.Book;
+import br.com.ada.adabook.dto.author.AuthorDetailsDTO;
 import br.com.ada.adabook.dto.author.AuthorListDTO;
 import br.com.ada.adabook.dto.author.AuthorSaveDTO;
+import br.com.ada.adabook.exceptions.AuthorHasBookException;
 import br.com.ada.adabook.exceptions.AuthorNotFoundException;
 import br.com.ada.adabook.mapper.AuthorMapper;
 import br.com.ada.adabook.repository.AuthorRepository;
+import br.com.ada.adabook.repository.BookRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -18,6 +21,7 @@ public class AuthorServiceImpl implements AuthorService {
 
     private final AuthorRepository authorRepository;
     private final AuthorMapper authorMapper;
+    private final BookRepository bookRepository;
 
     @Override
     public List<AuthorListDTO> list() {
@@ -26,7 +30,7 @@ public class AuthorServiceImpl implements AuthorService {
     }
 
     @Override
-    public AuthorDescriptionDTO findById(Long id) {
+    public AuthorDetailsDTO findById(Long id) {
         Author author = authorRepository.findById(id).orElseThrow(AuthorNotFoundException::new);
         return authorMapper.toAuthorDescriptionDTO(author);
     }
@@ -53,6 +57,12 @@ public class AuthorServiceImpl implements AuthorService {
         if (!authorRepository.existsById(id)) {
             throw new AuthorNotFoundException();
         }
-            authorRepository.deleteById(id);
+
+        Author author = authorRepository.findById(id).orElseThrow(AuthorNotFoundException::new);
+        if (author.getBooks() != null) {
+            throw new AuthorHasBookException();
+        }
+
+        authorRepository.deleteById(id);
     }
 }
